// src/app/interfaces/ICar.ts
export interface ICar {
    kenteken: string;
    voertuigsoort: string;
    merk: string;
    handelsbenaming: string;
    vervaldatum_apk: number;
    datum_tenaamstelling: number;
    bruto_bpm: number;
    inrichting: number;
    aantal_zitplaatsen: number;
    eerste_kleur: string;
    tweede_kleur: string;
    aantal_cilinders: number;
    cilinderinhoud: number;
    massa_ledig_voertuig: number;
    toegestane_maximum_massa_voertuig: number;
    maximum_massa_trekken_ongeremd: number;
    maximum_trekken_massa_geremd: number;
    retrofit_roetfilter: string;
    zuinigheidslabel: string;
    datum_eerste_toelating: number;
    datum_eerste_afgifte_nederland: number;
    wacht_op_keuren: string;
    catalogusprijs: number;
    wam_verzekerd: string;
    maximale_constructiesnelheid_brom_snorfiets: number;
    laadvermogen: number;
    oplegger_geremd: number;
    aanhangwagen_autonoom_geremd: number;
    aanhangwagen_middenas_geremd: number;
    vermogen_brom_snorfiets: number;
    aantal_staanplaatsen: number;
    aantal_deuren: number;
    aantal_wielen: number;
    afstand_hart_koppeling_tot_achterzijde_voertuig: number;
    afstand_voorzijde_voertuig_tot_hart_koppeling: number;
    afwijkende_maximum_snelheid: number;
    lengte: number;
    breedte: number;
    europese_voertuigcategorie: string;
    europese_voertuigcategorie_toevoeging: string;
    europese_uitvoeringcategorie_toevoeging: string;
    plaats_chassisnummer: string;
    technische_max_massa_voertuig: number;
    type: string;
    type_gasinstallatie: string;
    typegoedkeuringsnummer: string;
    variant: string;
    uitvoering: string;
    volgnummer_wijziging_eu_typegoedkeuring: number;
    vermogen_massarijklaar: number;
    wielbasis: number;
    export_indicator: string;
    openstaande_terugroepactie_indicator: string;
    vervaldatum_tachograaf: number;
    taxi_indicator: string;
    maximum_massa_samenstelling: number;
    aantal_rolstoelplaatsen: number;
    maximum_ondersteunende_snelheid: number;
}