// src/app/interfaces/ICarFuel.ts
export interface ICarFuel {
    kenteken: string;
    brandstof_volgnummer: string;
    brandstof_omschrijving: string;
    brandstofverbruik_buiten: string;
    brandstofverbruik_gecombineerd: string;
    brandstofverbruik_stad: string;
    co2_uitstoot_gecombineerd: string;
    co2_uitstoot_gewogen: string;
    geluidsniveau_rijdend: string;
    geluidsniveau_stationair: string;
    emissiecode_omschrijving: string;
    milieuklasse_eg_goedkeuring_licht: string;
    milieuklasse_eg_goedkeuring_zwaar: string;
    uitstoot_deeltjes_licht: string;
    uitstoot_deeltjes_zwaar: string;
    nettomaximumvermogen: string;
    nominaal_continu_maximumvermogen: string;
    roetuitstoot: string;
    toerental_geluidsniveau: string;
}